fixing delay navigation
