--- samrtble/samrtble/Assets.xcassets/gradient_Up.imageset/Clip path group.pdf
@@ -1,275 +1,0 @@
-%PDF-1.7
-
-1 0 obj
-  << /Length 2 0 R
-     /Range [ 0.000000 1.000000 0.000000 1.000000 0.000000 1.000000 ]
-     /Domain [ 0.000000 1.000000 ]
-     /FunctionType 4
-  >>
-stream
-{  0.203922 exch 0.752941 exch 0.768627 exch dup 0.000000 gt { exch pop exch pop exch pop dup 0.000000 sub 0.717647 mul 0.203922 add exch dup 0.000000 sub -0.082353 mul 0.752941 add exch dup 0.000000 sub -0.592157 mul 0.768627 add exch } if dup 1.000000 gt { exch pop exch pop exch pop 0.921569 exch 0.670588 exch 0.176471 exch } if pop }
-endstream
-endobj
-
-2 0 obj
-  338
-endobj
-
-3 0 obj
-  << /Type /XObject
-     /Length 4 0 R
-     /Group << /Type /Group
-               /S /Transparency
-               /I true
-            >>
-     /Subtype /Form
-     /Resources << /Pattern << /P1 << /Matrix [ 21.637260 188.147186 -188.147186 21.637260 777.781372 -88.606316 ]
-                                      /Shading << /Coords [ 0.000000 0.000000 1.000000 0.000000 ]
-                                                  /ColorSpace /DeviceRGB
-                                                  /Function 1 0 R
-                                                  /Domain [ 0.000000 1.000000 ]
-                                                  /ShadingType 2
-                                                  /Extend [ true true ]
-                                               >>
-                                      /PatternType 2
-                                      /Type /Pattern
-                                   >> >> >>
-     /BBox [ 0.000000 0.000000 375.000000 155.000000 ]
-  >>
-stream
-q
--1.000000 -0.000000 0.000000 -1.000000 597.377441 233.831726 cm
-/Pattern cs
-/P1 scn
-983.208374 26.521698 m
-978.208374 36.505692 966.102356 44.152710 953.122375 50.444702 c
-945.911377 53.944702 938.284424 57.087708 930.647400 60.208710 c
-923.570374 63.100708 909.198364 66.919708 909.773376 73.108704 c
-910.135376 77.014709 913.873413 79.488708 918.796387 81.121704 c
-919.099487 80.423538 919.366516 79.710281 919.596375 78.984711 c
-919.763367 78.451706 921.077393 78.646683 920.919373 79.179688 c
-920.682007 79.975113 920.399536 80.756409 920.073364 81.519714 c
-925.425232 82.898560 930.887939 83.804382 936.398376 84.226715 c
-938.570374 84.443710 937.777405 86.508698 935.609375 86.292694 c
-930.024353 85.888779 924.485657 84.994324 919.057373 83.619690 c
-913.785400 93.294693 901.271362 101.077713 886.003357 105.014709 c
-877.369873 107.234787 868.438965 108.071182 859.543396 107.492706 c
-848.170410 106.720703 837.804382 103.410706 828.007385 99.945709 c
-817.486389 96.223709 806.992371 91.956711 795.081360 90.165710 c
-785.148926 88.721344 775.050659 88.856339 765.160400 90.565704 c
-755.343384 92.265701 746.080383 95.540710 739.495361 100.428711 c
-731.430359 106.415710 729.032410 114.228683 730.559387 121.847687 c
-733.659363 137.310684 748.199402 150.928711 748.726379 166.747711 c
-749.161377 179.793701 742.180420 192.721710 730.389404 203.453705 c
-703.664429 227.776703 655.256409 237.184708 609.610413 232.159698 c
-562.583435 226.982697 523.789429 207.059708 499.083405 182.319702 c
-485.723419 168.943695 478.255432 154.055710 467.448425 139.919708 c
-456.641418 125.783707 440.677429 114.068703 418.021423 106.858704 c
-395.521423 99.703705 369.450409 97.975693 345.016418 102.004700 c
-333.358276 103.849472 321.979614 107.158638 311.150391 111.853699 c
-301.040405 116.339699 293.015411 122.683708 281.894409 126.281708 c
-259.036407 133.681702 232.902405 127.218697 209.158401 125.581696 c
-178.277405 123.453697 193.177414 146.020721 194.690414 156.519714 c
-198.187408 180.858719 171.377411 205.768692 133.909409 213.231689 c
-94.394409 221.103699 49.424408 206.943710 31.474409 184.362701 c
-30.391302 183.001312 29.399786 181.569473 28.506407 180.076691 c
-26.428408 181.244003 24.268787 182.259644 22.044409 183.115692 c
-15.329415 185.611786 8.187028 186.755066 1.028411 186.479706 c
--0.286589 186.441711 -0.371594 185.165695 0.953406 185.201691 c
-10.305410 185.573151 19.571272 183.287216 27.677412 178.608704 c
-25.769407 175.100830 24.529064 171.269272 24.019407 167.308716 c
-23.895407 166.322723 26.242413 166.308685 26.455412 167.253693 c
-27.245497 170.657562 28.542648 173.923386 30.303411 176.941711 c
-34.580086 174.041779 38.263149 170.351227 41.154411 166.068695 c
-49.903412 152.868698 45.501411 138.742706 41.278412 125.217712 c
-37.342411 112.617714 34.138409 100.068695 38.142410 87.304688 c
-41.654411 76.104691 49.675411 65.563690 60.342411 56.318695 c
-112.899414 10.772690 210.755402 6.292694 293.423401 6.228699 c
-392.574402 6.151703 491.160400 13.841690 590.171387 15.947693 c
-639.804382 17.002686 689.427368 16.471695 738.839355 13.228699 c
-787.646362 10.028702 836.250366 5.022705 885.393372 4.211700 c
-908.830627 3.715408 932.277344 4.531693 955.623352 6.656708 c
-963.772339 7.438705 975.993347 7.662704 981.833374 11.893707 c
-987.300354 15.857712 985.477356 21.989700 983.208374 26.521698 c
-h
-f
-n
-Q
-
-endstream
-endobj
-
-4 0 obj
-  3330
-endobj
-
-5 0 obj
-  << /Type /XObject
-     /Length 6 0 R
-     /Group << /Type /Group
-               /S /Transparency
-               /I true
-            >>
-     /Subtype /Form
-     /Resources << /XObject << /X1 3 0 R >>
-                   /ExtGState << /E2 << /ca 1.000000 >>
-                                 /E1 << /BM /Multiply >>
-                              >>
-                >>
-     /BBox [ 0.000000 0.000000 375.000000 155.000000 ]
-  >>
-stream
-q
-/E1 gs
-/E2 gs
-/X1 Do
-Q
-
-endstream
-endobj
-
-6 0 obj
-  25
-endobj
-
-7 0 obj
-  << /Type /XObject
-     /Length 8 0 R
-     /Group << /Type /Group
-               /S /Transparency
-            >>
-     /Subtype /Form
-     /Resources << /XObject << /X1 5 0 R >>
-                   /ExtGState << /E2 << /ca 0.300000 >>
-                                 /E1 << /BM /Multiply >>
-                              >>
-                >>
-     /BBox [ 0.000000 0.000000 375.000000 155.000000 ]
-  >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-/E1 gs
-/E2 gs
-/X1 Do
-Q
-
-endstream
-endobj
-
-8 0 obj
-  53
-endobj
-
-9 0 obj
-  << /Type /XObject
-     /Length 10 0 R
-     /Group << /Type /Group
-               /S /Transparency
-            >>
-     /Subtype /Form
-     /Resources << >>
-     /BBox [ 0.000000 0.000000 375.000000 155.000000 ]
-  >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
--1.000000 -0.000000 0.000000 -1.000000 377.000000 155.000000 cm
-0.000000 0.000000 0.000000 scn
-378.000000 155.000000 m
-0.000000 155.000000 l
-0.000000 0.000000 l
-378.000000 0.000000 l
-378.000000 155.000000 l
-h
-f
-n
-Q
-
-endstream
-endobj
-
-10 0 obj
-  245
-endobj
-
-11 0 obj
-  << /XObject << /X1 7 0 R >>
-     /ExtGState << /E1 << /SMask << /Type /Mask
-                                    /G 9 0 R
-                                    /S /Alpha
-                                 >>
-                          /Type /ExtGState
-                       >> >>
-  >>
-endobj
-
-12 0 obj
-  << /Length 13 0 R >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-/E1 gs
-/X1 Do
-Q
-
-endstream
-endobj
-
-13 0 obj
-  46
-endobj
-
-14 0 obj
-  << /Annots []
-     /Type /Page
-     /MediaBox [ 0.000000 0.000000 375.000000 155.000000 ]
-     /Resources 11 0 R
-     /Contents 12 0 R
-     /Parent 15 0 R
-  >>
-endobj
-
-15 0 obj
-  << /Kids [ 14 0 R ]
-     /Count 1
-     /Type /Pages
-  >>
-endobj
-
-16 0 obj
-  << /Pages 15 0 R
-     /Type /Catalog
-  >>
-endobj
-
-xref
-0 17
-0000000000 65535 f
-0000000010 00000 n
-0000000532 00000 n
-0000000554 00000 n
-0000004904 00000 n
-0000004927 00000 n
-0000005412 00000 n
-0000005433 00000 n
-0000005923 00000 n
-0000005944 00000 n
-0000006440 00000 n
-0000006463 00000 n
-0000006762 00000 n
-0000006866 00000 n
-0000006888 00000 n
-0000007067 00000 n
-0000007143 00000 n
-trailer
-<< /ID [ (some) (id) ]
-   /Root 16 0 R
-   /Size 17
->>
-startxref
-7204
-%%EOF
