+++ samrtble/samrtble/Assets.xcassets/gradient_Up.imageset/Contents.json
@@ -1,15 +1,23 @@
 {
   "images" : [
     {
-      "filename" : "Clip path group.pdf",
-      "idiom" : "universal"
+      "filename" : "Clip path group-4.png",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "Clip path group-5 1.png",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "Clip path group-6.png",
+      "idiom" : "universal",
+      "scale" : "3x"
     }
   ],
   "info" : {
     "author" : "xcode",
     "version" : 1
-  },
-  "properties" : {
-    "preserves-vector-representation" : true
   }
 }
