+++ samrtble/samrtble/Assets.xcassets/gradient_down.imageset/Contents.json
@@ -1,15 +1,23 @@
 {
   "images" : [
     {
-      "filename" : "Clip path group-2.pdf",
-      "idiom" : "universal"
+      "filename" : "Clip path group.png",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "Clip path group-2.png",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "Clip path group-3.png",
+      "idiom" : "universal",
+      "scale" : "3x"
     }
   ],
   "info" : {
     "author" : "xcode",
     "version" : 1
-  },
-  "properties" : {
-    "preserves-vector-representation" : true
   }
 }
