+++ samrtble/samrtble/VieweModels/RemoteNotificationModel.swift
@@ -13,7 +13,7 @@ class RemoteNotificationViewModel {
     static var FCMTOKEN_KEY = "fcm_token_key"
     
     func fetchRemoteNotification() {
-        if let phone = UserDefaults.standard.string(forKey: MainLoginViewController.PHONE_KEY){
+        if let phone = UserDefaults.standard.string(forKey: LoginViewController.PHONE_KEY){
             if let fcmToken = UserDefaults.standard.string(forKey: RemoteNotificationViewModel.FCMTOKEN_KEY){
                 let apiService = ApiService(phone: phone, type: "fcm-token")
                 apiService.fetchRemOteNotification(completeion: { (result) in
