+++ samrtble/samrtble/Viewes/Controller/ForgetPassword/ForgetPasswordViewController.xib
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
     <device id="retina6_12" orientation="portrait" appearance="light"/>
     <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21678"/>
         <capability name="Named colors" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -19,8 +20,13 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ForgetPasswordViewController" customModule="samrtble" customModuleProvider="target">
             <connections>
                 <outlet property="activityIndicatore" destination="8fi-2e-t83" id="HxR-u4-4Yy"/>
+                <outlet property="backButton" destination="P1f-L5-Ta2" id="znn-Ga-o8P"/>
                 <outlet property="emailTextField" destination="txL-NS-WaH" id="VDD-86-n7F"/>
                 <outlet property="emailView" destination="FcD-qy-ZNV" id="0hz-1U-Bii"/>
+                <outlet property="errorLabel" destination="FQU-kq-yag" id="H0K-Wm-1K6"/>
+                <outlet property="sendButton" destination="KD4-OK-i8x" id="8q6-Px-fVL"/>
+                <outlet property="subtitleLabel" destination="WDK-Pa-UVj" id="Zte-ZS-Zrg"/>
+                <outlet property="titleLabel" destination="bR2-aB-svk" id="fZG-71-Bii"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
@@ -94,7 +100,7 @@
                                         <constraint firstItem="F4F-nj-8ZG" firstAttribute="leading" secondItem="FcD-qy-ZNV" secondAttribute="leading" constant="16" id="yFu-Ar-5e6"/>
                                     </constraints>
                                 </view>
-                                <label hidden="YES" opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" insetsLayoutMarginsFromSafeArea="NO" text="This filed can't be empty" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="FQU-kq-yag">
+                                <label hidden="YES" opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" insetsLayoutMarginsFromSafeArea="NO" text="This filed can't be empty" textAlignment="center" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="FQU-kq-yag">
                                     <rect key="frame" x="0.0" y="0.0" width="361" height="24"/>
                                     <constraints>
                                         <constraint firstAttribute="height" constant="24" id="4f1-kc-DMy"/>
@@ -105,8 +111,22 @@
                                 </label>
                             </subviews>
                         </stackView>
+                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="P1f-L5-Ta2">
+                            <rect key="frame" x="0.0" y="64" width="361" height="30"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="30" id="YcR-rk-nVX"/>
+                            </constraints>
+                            <fontDescription key="fontDescription" name="MyriadPro-Bold" family="Myriad Pro" pointSize="16"/>
+                            <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
+                            <state key="normal" title="Back">
+                                <color key="titleColor" name="primary-dark"/>
+                            </state>
+                            <connections>
+                                <action selector="didTapBackButton:" destination="-1" eventType="touchUpInside" id="iwp-sC-aX8"/>
+                            </connections>
+                        </button>
                         <button opaque="NO" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="KD4-OK-i8x" customClass="GradientButton" customModule="samrtble" customModuleProvider="target">
-                            <rect key="frame" x="0.0" y="64" width="361" height="50"/>
+                            <rect key="frame" x="0.0" y="108" width="361" height="50"/>
                             <constraints>
                                 <constraint firstAttribute="height" constant="50" id="bKd-dE-zEe"/>
                             </constraints>
@@ -124,20 +144,6 @@
                             </userDefinedRuntimeAttributes>
                             <connections>
                                 <action selector="didTapSendButton:" destination="-1" eventType="touchUpInside" id="cGz-nw-7g2"/>
-                            </connections>
-                        </button>
-                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="P1f-L5-Ta2">
-                            <rect key="frame" x="0.0" y="128" width="361" height="30"/>
-                            <constraints>
-                                <constraint firstAttribute="height" constant="30" id="YcR-rk-nVX"/>
-                            </constraints>
-                            <fontDescription key="fontDescription" name="MyriadPro-Bold" family="Myriad Pro" pointSize="16"/>
-                            <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                            <state key="normal" title="Back">
-                                <color key="titleColor" name="primary-dark"/>
-                            </state>
-                            <connections>
-                                <action selector="didTapBackButton:" destination="-1" eventType="touchUpInside" id="iwp-sC-aX8"/>
                             </connections>
                         </button>
                     </subviews>
