+++ samrtble/samrtble/Viewes/Controller/Table/TableHomeViewController.swift
@@ -161,14 +161,11 @@ class TableHomeViewController: UIViewController {
         }
     }
     
-    func callApi(){
+    func callApi() {
         ActivityIndecatorBehaviour.activityIndecatorAction(activityIndecator: activityIndecator, status: .start)
-        
         let defaults = UserDefaults.standard
-        mobilePhone = defaults.string(forKey: MainLoginViewController.PHONE_KEY)
-        
+        mobilePhone = defaults.string(forKey: LoginViewController.PHONE_KEY)
         homaViewModel.fetchDataFromApi(phone: mobilePhone)
-        
     }
     
     @IBAction func schoolSegmentAction(_ sender: Any) {
@@ -281,10 +278,9 @@ extension TableHomeViewController : SideMenuNavigationControllerDelegate {
     }
     
     private func logout() {
-//        UserDefaults.standard.set(nil, forKey: MainLoginViewController.PHONE_KEY)
-//        UserDefaults.standard.set(nil, forKey: USER_TOKEN)
-
-        let asd = ForgetPasswordViewController()
-        navigationController?.setViewControllers([asd], animated: true)
+        UserDefaults.standard.set(nil, forKey: LoginViewController.PHONE_KEY)
+        UserDefaults.standard.set(nil, forKey: USER_TOKEN)
+        let vc = LoginViewController()
+        navigationController?.setViewControllers([vc], animated: true)
     }
 }
